+++ src/qutip_qip/circuit.py
@@ -49,7 +49,7 @@ from .operations import (
 )
 from .operations.gates import _gate_label
 from qutip import basis, ket2dm, qeye
-from qutip.qobj import Qobj
+from qutip import Qobj
 from qutip.measurement import measurement_statistics
 
 
@@ -2130,3 +2130,30 @@ class CircuitSimulator:
         else:
             raise NotImplementedError(
                 "mode {} is not available.".format(self.mode))
+
+# For Decomposition functions
+def decomposed_gates_to_circuit(decomposed_gate,num_qubits):
+    """This function takes the input from a decomposition function and returns
+    a quantum circuit.
+    """
+    # as decomposed_gate contains information about targets/control, there's no
+    # additional input of target here.
+    # In addition, there's no check if the gates are valid for number of qubits
+    # because this is done in a decomposition function before output.
+    if isinstance(decomposed_gate,tuple) == True:
+        q_circuit = QubitCircuit(num_qubits, reverse_states=False)
+        for i in decomposed_gate:
+            q_circuit.add_gate(i)
+        return(q_circuit)
+    else:
+        raise TypeError("Input is not a tuple of gates.")
+
+def compute_unitary(quantum_circuit):
+    """Evaluates all the gates in the quantum circuit.
+    """
+    if isinstance(quantum_circuit, QubitCircuit) == True:
+        gate_list = quantum_circuit.propagators()
+        matrix_of_all_gates = gate_sequence_product(gate_list)
+        return(matrix_of_all_gates)
+    else:
+        raise TypeError("Input is not of type QubitCircuit.")
