+++ src/qutip_qip/decompose/__init__.py
@@ -1,0 +1,1 @@
+from qutip_qip.decompose.single_decompositions import (ABC_decomposition, decompose_to_rotation_matrices)
