+++ src/qutip_qip/decompose/single_decompositions.py
@@ -1,21 +1,24 @@
 import numpy as np
 import cmath
 
-from qutip.qobj import Qobj
-from qutip_qip.decompositions.general_decompositions import (check_input, check_input_shape,
+from qutip import Qobj
+from qutip_qip.decompose._utility import (check_input, check_input_shape,
 convert_qobj_gate_to_array, extract_global_phase, MethodError, GateError)
 
-from qutip_qip.operations import *
 from qutip_qip.circuit import QubitCircuit, Gate
 
 
 # Functions for decompose_to_rotation_matrices
-single_qubit = 1
-def _angles_for_ZYZ(input_gate, single_qubit):
+def _angles_for_ZYZ(input_gate, num_qubits=1):
     """ Finds and returns the angles for ZYZ rotation matrix. These are
     used to change ZYZ to other combinations.
+
+    Parameters
+    ----------
+    input_gate : :class:`qutip.Qobj`
+        The gate matrix that's supposed to be decomposed should be a Qobj.
     """
-    global_phase_angle = extract_global_phase(input_gate,single_qubit)
+    global_phase_angle = extract_global_phase(input_gate,num_qubits)
     input_array = convert_qobj_gate_to_array(input_gate)
     input_array = input_array/cmath.rect(1,global_phase_angle)
     # separate all the elements
@@ -41,7 +44,7 @@ def _ZYZ_rotation(input_gate, num_of_qubits, target):
     input_gate : :class:`qutip.Qobj`
         The matrix that's supposed to be decomposed should be a Qobj.
     """
-    angle_list = _angles_for_ZYZ(input_gate, single_qubit)
+    angle_list = _angles_for_ZYZ(input_gate, 1)
     alpha = angle_list[0]
     beta = angle_list[2]
     theta = angle_list[1]
@@ -69,7 +72,7 @@ def _ZXZ_rotation(input_gate, num_of_qubits, target):
     input_gate : :class:`qutip.Qobj`
         The matrix that's supposed to be decomposed should be a Qobj.
     """
-    angle_list = _angles_for_ZYZ(input_gate, single_qubit)
+    angle_list = _angles_for_ZYZ(input_gate, 1)
     alpha = angle_list[0]
     alpha = alpha - np.pi/2
     beta = angle_list[2]
@@ -93,9 +96,9 @@ def _ZXZ_rotation(input_gate, num_of_qubits, target):
 
 _rotation_matrices_dictionary ={"ZYZ": _ZYZ_rotation,
                                 "ZXZ": _ZXZ_rotation,
-                                }
+                                } # other combinations to add here
 
-def decompose_to_rotation_matrices(input_gate, num_of_qubits, target, method):
+def decompose_to_rotation_matrices(input_gate, method, num_of_qubits, target=0):
     r""" An input 1-qubit gate is expressed as a product of rotation matrices
     :math:`\textrm{R}_i` and :math:`\textrm{R}_j`.
 
@@ -157,7 +160,8 @@ def decompose_to_rotation_matrices(input_gate, num_of_qubits, target, method):
     else:
         raise MethodError("Invalid method chosen.")
 
-
+# new combinations need to be added by creating a dictionary for these.
+# The fucntion below will decompose the amtrix as a product of Rz, Ry and Pauli X only.
 def ABC_decomposition(input_gate, num_of_qubits, target):
     r""" An input 1-qubit gate is expressed as a product of rotation matrices
     :math:`\textrm{R}_z`, :math:`\textrm{R}_y` and Pauli :math:`\textrm{X}`.
@@ -195,7 +199,7 @@ def ABC_decomposition(input_gate, num_of_qubits, target):
             raise GateError("This method is valid for single qubit gates only. Provide a target qubit for larger circuits. ")
 
 
-    global_phase_angle = extract_global_phase(input_gate,single_qubit)
+    global_phase_angle = extract_global_phase(input_gate,1)
     input_array = convert_qobj_gate_to_array(input_gate)
     input_array = input_array/cmath.rect(1,global_phase_angle)
     # separate all the elements
