--- src/qutip_qip/decompositions/__init__.py
@@ -1,7 +1,0 @@
-from qutip_qip.decompositions.general_decompositions import (check_input,
-check_input_shape, convert_qobj_gate_to_array, extract_global_phase)
-
-from qutip_qip.decompositions.single_decompositions import (_ZYZ_rotation, _ZXZ_rotation, _rotation_matrices_dictionary,
-                                                            ABC_decomposition, decompose_to_rotation_matrices)
-
-from qutip_qip.decompositions.decompositions_extras import (decomposed_gates_to_circuit, matrix_of_decomposed_gates)
