+++ tests/gate_decompose/test_single_decompositions.py
@@ -2,14 +2,12 @@ import numpy as np
 import cmath
 import pytest
 
-from qutip.qobj import Qobj
-from qutip.metrics import average_gate_fidelity
+from qutip import Qobj, average_gate_fidelity
 
-
-from qutip_qip.decompositions.single_decompositions import (_ZYZ_rotation, _ZXZ_rotation,
+from qutip_qip.decompose.single_decompositions import (_ZYZ_rotation, _ZXZ_rotation,
                                                             ABC_decomposition, decompose_to_rotation_matrices)
 
-from qutip_qip.decompositions.decompositions_extras import (decomposed_gates_to_circuit, matrix_of_decomposed_gates)
+from qutip_qip.circuit import (decomposed_gates_to_circuit, compute_unitary)
 
 # Fidelity closer to 1 means the two states are similar to each other
 H = Qobj([[1/np.sqrt(2),1/np.sqrt(2)],[1/np.sqrt(2),-1/np.sqrt(2)]])
@@ -27,7 +25,7 @@ def test_single_qubit_to_rotations(gate, method):
     target = 0
     gate_list = method(gate,1,target)
     decomposed_gates_circuit = decomposed_gates_to_circuit(gate_list,1)
-    decomposed_gates_final_matrix = matrix_of_decomposed_gates(decomposed_gates_circuit)
+    decomposed_gates_final_matrix = compute_unitary(decomposed_gates_circuit)
     fidelity_of_input_output = average_gate_fidelity(gate, decomposed_gates_final_matrix)
     assert(np.isclose(fidelity_of_input_output,1.0))
 
@@ -38,7 +36,7 @@ def test_check_single_qubit_to_decompose_to_rotations(gate, method):
     target = 0
     gate_list = decompose_to_rotation_matrices(gate,1,target,method)
     decomposed_gates_circuit = decomposed_gates_to_circuit(gate_list,1)
-    decomposed_gates_final_matrix = matrix_of_decomposed_gates(decomposed_gates_circuit)
+    decomposed_gates_final_matrix = compute_unitary(decomposed_gates_circuit)
     fidelity_of_input_output = average_gate_fidelity(gate, decomposed_gates_final_matrix)
     assert(np.isclose(fidelity_of_input_output,1.0))
