+++ tests/test_circuit.py
@@ -38,7 +38,7 @@ from pathlib import Path
 
 
 from qutip_qip.circuit import (
-    QubitCircuit, CircuitSimulator, Measurement)
+    QubitCircuit, CircuitSimulator, Measurement, compute_unitary, decomposed_gates_to_circuit)
 from qutip import (tensor, Qobj, ptrace, rand_ket, fock_dm, basis,
                    rand_dm, bell_state, ket2dm, identity)
 from qutip_qip.qasm import read_qasm
@@ -47,6 +47,8 @@ from qutip_qip.operations import (
     _ctrl_gates, _single_qubit_gates, _swap_like, _toffoli_like, _fredkin_like,
     _para_gates
 )
+
+from qutip_qip.decompose.single_decompositions import _ZYZ_rotation
 
 import qutip as qp
 
@@ -674,3 +676,42 @@ class TestQubitCircuit:
             r"  \gate{{\rm H}}  &  \meter &  \qw  &  \qw  &  \qw  & \qw \\ ",
             "",
         ])
+
+# Tests for decomposed_gates_to_circuit
+@pytest.mark.parametrize("invalid_input",[np.array([[1,1],[1,1]]),([[1,1],[1,1]]),1.5,3])
+def test_decomposed_gates_to_circuit(invalid_input):
+    """Checks if correct error is raised when input is anything but a tuple of gates.
+    """
+    with pytest.raises(TypeError,match="Input is not a tuple of gates."):
+        decomposed_gates_to_circuit(invalid_input,1)
+
+H = Qobj([[1/np.sqrt(2),1/np.sqrt(2)],[1/np.sqrt(2),-1/np.sqrt(2)]])
+sigmax = Qobj([[0,1],[1,0]])
+H_zyz_gates = _ZYZ_rotation(H, 1, 0)
+sigmax_zyz_gates = _ZYZ_rotation(sigmax, 1, 0)
+
+@pytest.mark.parametrize("valid_input",[H_zyz_gates,sigmax_zyz_gates])
+def test_decomposed_gates_to_circuit(valid_input):
+    """Checks if output is of type QubitCircuit.
+    """
+    assert(isinstance(decomposed_gates_to_circuit(valid_input,1),QubitCircuit))
+
+
+H_zyz_quantum_circuit = decomposed_gates_to_circuit(H_zyz_gates, 1)
+sigmax_zyz_quantum_circuit = decomposed_gates_to_circuit(sigmax_zyz_gates, 1)
+sigmax_zyz_output = (sigmax_zyz_quantum_circuit)
+# Tests for matrix_of_decomposed_gates
+@pytest.mark.parametrize("invalid_input",[np.array([[1,1],[1,1]]),([[1,1],[1,1]]),1.5,3])
+def test_matrix_of_decomposed_gates(invalid_input):
+    """Checks if correct error is raised when input is anything but a quantum circuit.
+    """
+    with pytest.raises(TypeError,match="Input is not of type QubitCircuit."):
+        compute_unitary(invalid_input)
+
+
+@pytest.mark.parametrize("valid_input",[H_zyz_quantum_circuit,sigmax_zyz_quantum_circuit])
+def test_matrix_of_decomposed_gates(valid_input):
+    """Checks if final output is a Qobj.
+    """
+    final_output=compute_unitary(valid_input)
+    assert(isinstance(final_output, Qobj))
